requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysql-connector-python
alembic
python-jose[cryptography]
